simplify models import & extract ConstantModel
